Update Session, Session Authentication, move Abstract Session from Session to Authentication folders
